make hivers' list a component
