+++ app/components/hivers/template.html
@@ -1,4 +1,4 @@
-<div ng-controller="HiversCtrl">
+<div class="container">
   <h4>Hivers</h4>
   <span class="list-group">
     <span class="col-sm-3 ph-sm-padding"
