+++ app/controllers/index.ts
@@ -1,6 +1,5 @@
 import documentText from './documentText';
 import authReturn from './authReturn';
-import hivers from './hivers';
 import meta from './meta';
 import jobs from './jobs';
 import passwordRequest from './passwordRequest';
@@ -8,7 +7,6 @@ import passwordRequest from './passwordRequest';
 export default function(app) {
   documentText(app);
   authReturn(app);
-  hivers(app);
   meta(app);
   jobs(app);
   passwordRequest(app);
